refactor: 알람 useHooks 구조로 교체
