+++ src/components/alarm/AlarmPopup/index.jsx
@@ -1,35 +1,14 @@
-import useClickAway from '../../../hooks/useClickAway';
 import ReactDOM from 'react-dom';
-import { useMemo, useEffect } from 'react';
+import useAlarm from '../../../hooks/useAlarm';
+import useClickAway from '../../../hooks/useClickAway';
 import AlarmPopupItem from '../AlarmPopupItem';
-import { AlarmPopupContainer, Title, AlarmList, AlarmNoItem } from './style';
-import { useSetRecoilState } from 'recoil';
-import { selectedPostState } from '../../../recoil/postStates';
-import { useNavigate } from 'react-router-dom';
+import { AlarmList, AlarmNoItem, AlarmPopupContainer, Title } from './style';
 
 const AlarmPopup = ({ visible = false, onClose, target, alarms }) => {
-  const navigate = useNavigate();
-  const setPostId = useSetRecoilState(selectedPostState);
-
+  const { element, handleClickAlarm } = useAlarm(target);
   const ref = useClickAway(() => {
     onClose && onClose();
   });
-
-  const element = useMemo(() => document.createElement('div'), []);
-  useEffect(() => {
-    if (target) {
-      target.appendChild(element);
-      return () => {
-        target.removeChild(element);
-      };
-    }
-  });
-
-  const handleClickAlarm = (postId) => {
-    setPostId(postId);
-    navigate(`/p/${postId}`);
-    onClose();
-  };
 
   return ReactDOM.createPortal(
     <AlarmPopupContainer ref={ref} style={{ display: visible ? 'block' : 'none' }}>
